Refactoring environment variables validation
- Splitting Zod validators for env vars in separate files (in back, play and map-storage)
- Adding a small unit test
- Putting utility functions in shared-lib
