+++ back/tests/EnvironmentVariableValidator.test.ts
@@ -1,0 +1,24 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+
+import { describe, expect, it } from "vitest";
+import { EnvironmentVariables } from "../src/Enum/EnvironmentVariableValidator";
+
+describe("EnvironmentVariable", () => {
+    it("should validate properly URLs", () => {
+        let result = EnvironmentVariables.safeParse({
+            PLAY_URL: "https://exemple.com",
+        });
+        expect(result.success).toBe(true);
+
+        result = EnvironmentVariables.safeParse({
+            PLAY_URL: "https://12.12.12.12",
+        });
+        expect(result.success).toBe(true);
+
+        result = EnvironmentVariables.safeParse({
+            PLAY_URL: "https://12.12.12.12",
+            ADMIN_API_URL: "",
+        });
+        expect(result.success).toBe(true);
+    });
+});
