+++ libs/shared-utils/package.json
@@ -19,7 +19,8 @@
   "license": "ISC",
   "dependencies": {
     "@workadventure/map-editor": "1.0.0",
-    "@workadventure/tiled-map-type-guard": "^2.1.0"
+    "@workadventure/tiled-map-type-guard": "^2.1.0",
+    "zod": "^3.19.1"
   },
   "devDependencies": {
     "@jest/globals": "^29.3.1",
