+++ libs/shared-utils/src/index.ts
@@ -1,3 +1,4 @@
 export * as Color from "./String/color";
 export * as ShortHash from "./String/shortHash";
 export * as Jitsi from "./Jitsi/slugify";
+export * as EnvironmentVariableUtils from "./EnvironmentVariables/EnvironmentVariableUtils";
